requirements.txt
# --------- pytorch --------- #
torch==1.13.1
torchvision==0.14.1
pytorch-lightning==1.8.1
torchmetrics==0.11.1

# --------- hydra --------- #
hydra-core==1.2.0
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# --------- loggers --------- #
wandb==0.13.5
# neptune-client
# mlflow
comet-ml==3.31.19

# --------- others --------- #
pyrootutils~=1.0.4
pre-commit      # hooks for applying linters on commit
rich~=12.6.0
pytest~=7.2.0
omegaconf~=2.2.3
matplotlib~=3.6.2
numpy~=1.23.5
pandas~=1.5.2
pillow~=9.3.0
packaging~=21.3
setuptools~=65.5.0
scikit-learn~=1.1.3
torch-intermediate-layer-getter==0.1.post1
retinaface-pytorch==0.0.7
ultralytics==8.0.53
timm==0.6.12
seaborn==0.12.2
